Move cleanup method to cleanup utils
