+++ Assets/Tests/BeatmapActionTest.cs
@@ -21,7 +21,7 @@ namespace Tests
         public void ContainerCleanup()
         {
             BeatmapActionContainer.RemoveAllActionsOfType<BeatmapAction>();
-            TestUtils.CleanupNotes();
+            CleanupUtils.CleanupNotes();
             TestUtils.ReturnSettings();
         }
