+++ Assets/Tests/MirrorTest.cs
@@ -63,7 +63,7 @@ namespace Tests
         public void ContainerCleanup()
         {
             BeatmapActionContainer.RemoveAllActionsOfType<BeatmapAction>();
-            TestUtils.CleanupNotes();
+            CleanupUtils.CleanupNotes();
             TestUtils.ReturnSettings();
         }
