+++ Assets/Tests/StrobeGeneratorTest.cs
@@ -23,7 +23,7 @@ namespace Tests
         public void ContainerCleanup()
         {
             BeatmapActionContainer.RemoveAllActionsOfType<BeatmapAction>();
-            TestUtils.CleanupEvents();
+            CleanupUtils.CleanupEvents();
             TestUtils.ReturnSettings();
         }
